+++ app/auth/anonymous/__init__.py
@@ -1,0 +1,5 @@
+from ..models import User
+
+
+def get_user():
+    return User(id="n/a", name="Anonymous")
