+++ tests/test_auth.py
@@ -8,7 +8,7 @@ from app.auth.entraid import get_key_set
 @pytest.mark.anyio
 async def test_custom_get_jwks_json():
     with patch(
-        "app.auth.jwks.get_jwks_json", new_callable=AsyncMock
+        "app.auth.entraid.jwks.get_jwks_json", new_callable=AsyncMock
     ) as mock_get_jwks_json:
         mock_get_jwks_json.return_value = {
             "keys": [{"kty": "RSA", "n": "test", "e": "AQAB"}]
